add LogoutButton component
