--- bookstore-client/src/context/AuthContext.tsx
@@ -1,42 +1,0 @@
-import React, { createContext, useContext, useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { useNavigate } from 'react-router-dom';
-import { RootState } from '../store/store';
-import { login as loginAction, logout as logoutAction } from '../store/authSlice';
-
-interface AuthContextType {
-    isAuthenticated: boolean;
-    login: () => void;
-    logout: () => void;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export const useAuth = () => {
-    const context = useContext(AuthContext);
-    if (!context) {
-        throw new Error('useAuth must be used within an AuthProvider');
-    }
-    return context;
-};
-
-export const AuthProvider: React.FC<any> = ({ children }) => {
-    const isAuthenticated = useSelector((state: RootState) => state.auth.isAuthenticated);
-    const dispatch = useDispatch();
-    const navigate = useNavigate();
-
-    useEffect(() => {
-        if (!isAuthenticated) {
-            navigate('/auth');
-        }
-    }, [isAuthenticated, navigate]);
-
-    const login = () => dispatch(loginAction());
-    const logout = () => dispatch(logoutAction());
-
-    return (
-        <AuthContext.Provider value={{ isAuthenticated, login, logout }}>
-            {children}
-        </AuthContext.Provider>
-    );
-};
