+++ bookstore-client/src/pages/AuthPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { RootState } from '../store/store';
+import { RootState } from '../redux/store';
 import AuthDialog from '../components/AuthDialog';
 
 const AuthPage: React.FC = () => {
